- Changed uint32_t to size_t for calls to  GetBufferPointerUnsafe as this inconsistency caused  some compilers to fail.

+++ src/server/npcmanager.cpp
@@ -734,7 +734,7 @@ void NPCManager::HandleCommandList(MsgEntry *me)
             case psNPCCommandsMessage::CMD_DRDATA:
             {
                 // extract the data
-                uint32_t len = 0;
+                size_t len = 0;
 				void *data = list.msg->GetBufferPointerUnsafe(len);
 
                 // Make sure we haven't run past the end of the buffer
